requirements.txt
asgiref
Django
crispy-bootstrap5
django-ckeditor
django-js-asset
Pillow
sqlparse
tzdata
django-crispy-forms
